# Backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/inventory": {
      "get": {
        "operationId": "InventoryController_getInventory_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DistributorInventory"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Inventory",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/v1/inventory/adjust": {
      "post": {
        "operationId": "InventoryController_adjustManualStock_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adjust Manual Stock",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/v1/inventory/{id}/movements": {
      "get": {
        "operationId": "InventoryController_getMovements_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Movements",
        "tags": [
          "Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vikas Inventory RN Platform API",
    "description": "Enterprise Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "DistributorInventory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id"
          },
          "distributor_id": {
            "type": "string",
            "description": "Distributor id"
          },
          "product_id": {
            "type": "string",
            "description": "Product id"
          },
          "available_quantity": {
            "type": "number",
            "description": "Available quantity"
          },
          "reserved_quantity": {
            "type": "number",
            "description": "Reserved quantity"
          },
          "backordered_quantity": {
            "type": "number",
            "description": "Backordered quantity"
          },
          "low_stock_threshold": {
            "type": "number",
            "description": "Low stock threshold"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at"
          }
        },
        "required": [
          "id",
          "distributor_id",
          "product_id",
          "available_quantity",
          "reserved_quantity",
          "backordered_quantity",
          "low_stock_threshold",
          "created_at",
          "updated_at"
        ]
      },
      "AdjustInventoryDto": {
        "type": "object",
        "properties": {
          "distributor_id": {
            "type": "string",
            "description": "Distributor id"
          },
          "manufacturer_id": {
            "type": "string",
            "description": "Manufacturer id"
          },
          "product_id": {
            "type": "string",
            "description": "Product id"
          },
          "movement_type": {
            "type": "string",
            "description": "Movement type"
          },
          "quantity_change": {
            "type": "number",
            "description": "Quantity change"
          },
          "reason": {
            "type": "string",
            "description": "Reason"
          }
        },
        "required": [
          "product_id",
          "movement_type",
          "quantity_change"
        ]
      },
...
    }
  }
}